CHEF-22096 : model changes
